Working Build as per requirements (#1)
* Working Build

* Working Year Filter

* Some cleaning
