+++ .env.dist
@@ -2,21 +2,24 @@
 # AIR_MODE=debug/release
 AIR_MODE="debug"
 
+# Which port to use for the API
+API_PORT="3000"
+
 # Should the App clean the DB on start
 # POSTGRES_DROP=true/false
 POSTGRES_DROP=true
 
 # Postgres connection info
-POSTGRES_HOST="host.docker.internal"
+POSTGRES_HOST="postgres"
 POSTGRES_PORT="5432"
-POSTGRES_USER="appuser"
-POSTGRES_PASSWORD="Password123!"
-POSTGRES_DB="tzktdelegator"
+POSTGRES_USERNAME="appuser"
+POSTGRES_PASSWORD="SecretsOfPg!5432"
+POSTGRES_DB="tzktdelegations"
 
 # PGAdmin connection info
 PGA_PORT="8080"
 PGA_USERNAME="admin@postgres.com"
-PGA_PASSWORD="Admin!123"
+PGA_PASSWORD="Password123!"
 
 # Hasura connection info
 HSR_PORT="8081"
