+++ .github/workflows/actions.yml
@@ -1,5 +1,5 @@
 # YAML file for Github Actions to use to build the Docker Image
-name: Docker Image
+name: Github Actions Build
 
 on:
   push:
@@ -12,5 +12,5 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-      - name: Build the Docker image
-        run: docker build .
+      - name: Build the Docker Compose
+        run: docker-compose build
