+++ Dockerfile.api
@@ -1,0 +1,13 @@
+FROM --platform=linux/amd64 golang:1.20.5
+
+WORKDIR /app/api
+
+RUN go install github.com/cosmtrek/air@latest
+
+COPY app/api/go.mod app/api/go.sum ./
+RUN go mod download
+
+COPY ./app/api/ ./
+
+# CMD ["air", "-c", ".air.toml"]
+CMD ["go", "run", "."]
