+++ app/indexer/go.mod
@@ -1,0 +1,5 @@
+module tzkt-delegations-indexer
+
+go 1.20
+
+require github.com/lib/pq v1.10.9
