+++ docs/ASSIGNEMENT.md
@@ -1,6 +1,8 @@
 # Exercise: Tezos Delegation Service
 
-In this exercise, you will build a service that gathers new [delegations](https://opentezos.com/baking/delegating/) made on the Tezos protocol and exposes them through a public API. 
+Build a service that :
+- gathers new [delegations](https://opentezos.com/baking/delegating/) made on the Tezos protocol,
+- exposes them through a public API. 
 
 ## Requirements:
