+++ postman/collection.json
@@ -1,0 +1,91 @@
+{
+	"info": {
+		"_postman_id": "936db90b-43d4-4cd7-ae8c-e92f10709233",
+		"name": "TZKT Delegations API",
+		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
+		"_exporter_id": "1823843"
+	},
+	"item": [
+		{
+			"name": "GetDelegationsCount",
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://api.tzkt.io/v1/operations/delegations/count",
+					"protocol": "https",
+					"host": [
+						"api",
+						"tzkt",
+						"io"
+					],
+					"path": [
+						"v1",
+						"operations",
+						"delegations",
+						"count"
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "/XTZ/Delegations",
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "http://localhost:3000/xtz/delegations?year=2019",
+					"protocol": "http",
+					"host": [
+						"localhost"
+					],
+					"port": "3000",
+					"path": [
+						"xtz",
+						"delegations"
+					],
+					"query": [
+						{
+							"key": "year",
+							"value": "2019"
+						}
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "GetDelegations",
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "https://api.tzkt.io/v1/operations/delegations?limit=10000&offset=0",
+					"protocol": "https",
+					"host": [
+						"api",
+						"tzkt",
+						"io"
+					],
+					"path": [
+						"v1",
+						"operations",
+						"delegations"
+					],
+					"query": [
+						{
+							"key": "limit",
+							"value": "10000"
+						},
+						{
+							"key": "offset",
+							"value": "0"
+						}
+					]
+				}
+			},
+			"response": []
+		}
+	]
+}
